+++ .gitignore
@@ -1,1 +1,6 @@
 public/
+myenv/
+mlruns/
+.ipynb_checkpoints/
+.DS_store
+.vscode/
