+++ content/data_activation/connect.md
@@ -1,13 +1,13 @@
 +++
-title = "Connect Databricks to Hightouch"
+title = "Connect Data Warehouse to Hightouch"
 weight = 1
 +++
 
-You can easily connect to Hightouch from Databricks using [Partner Connect](https://docs.databricks.com/integrations/partner-connect/reverse-etl.html).
+You can easily connect to Hightouch from your data warehouse using either [Databricks Partner Connect](https://docs.databricks.com/integrations/partner-connect/reverse-etl.html) or [Snowflake Partner Connect](https://hightouch.com/blog/hightouch-snowflake-partner-connect).
 
-1. Make sure your Databricks account, workspace, and the signed-in user all meet the [requirements](https://docs.databricks.com/integrations/partner-connect/index.html#requirements) for Partner Connect.
+1. Make sure your data warehouse account, workspace, and the signed-in user all meet the necessary requirements for Partner Connect.
 2. In the sidebar, click **Partner Connect**.
-3. Find the Hightouch tile. If the tile has a check mark icon, stop here, as your workspace is already connected. 
+3. Find the Hightouch tile. If the tile has a check mark icon, stop here, as your workspace is already connected.
 
 ![hightouch_partner_connect](../images/hightouch_partner_connect.png?width=30pc)
 
@@ -16,8 +16,8 @@ Otherwise, follow the on-screen directions to finish creating the connection.
 
 ![hightouch_partner_connect_instructions](../images/hightouch_partner_connect_instructions.png?width=50pc)
 
-Once set-up you should see your Databricks cluster under the Sources tab in Hightouch:
+Once set-up you should see your new connection under the Sources tab in Hightouch:
 
 ![hightouch_sources](../images/hightouch_sources.png?width=40pc)
 
-Alternatively you can [connect to Hightouch manually](https://docs.databricks.com/integrations/reverse-etl/hightouch.html#connect-to-hightouch-manually).
+Alternatively you can connect to Hightouch manually with [Databricks](https://docs.databricks.com/integrations/reverse-etl/hightouch.html#connect-to-hightouch-manually) or [Snowflake](https://hightouch.com/docs/sources/snowflake).
