+++ content/next_steps/summary.md
@@ -7,10 +7,10 @@ weight = 3
 
 Let's go over what you have built:
 
-1. From the [Advanced Analytics for Web](https://docs.snowplow.io/accelerators/web/) accelerator, you have created and modeled Snowplow behavioral data from your website, stored in your Databricks Lakehouse
-2. Collected user features and trained a propensity to convert machine learning model using Databricks managed MLflow
-3. Productionalized your ML model using Databricks model serving capability, saving your model output into a table of user propensity scores
-4. Connected your Databricks Lakehouse to Hightouch and built an audience using your user propensity scores
+1. From the [Advanced Analytics for Web](https://docs.snowplow.io/accelerators/web/) accelerator, you have created and modeled Snowplow behavioral data from your website, stored in your data warehouse
+2. Collected user features and trained a propensity to convert machine learning model
+3. Productionalized your ML model, saving your model output into a table of user propensity scores
+4. Connected your data warehouse to Hightouch and built an audience using your user propensity scores
 5. Synched and scheduled your audience to your chosen marketing tools and launched at least one ad campaign targeting high propensity users
 6. Built reporting on your ad campaign spend, highlighting the value of your ML model
