+++ content/predictive_ml_models/Databricks/model_training.md
@@ -1,0 +1,12 @@
++++
+title = "Model Training"
+weight = 1
++++
+
+#### Train predictive model using LightGBM and MLflow
+
+This propensity to convert solution provides a very flexible way to identify who among your audience is most likely to actually engage with you, for example request a demo / sign up for a trial, purchase a first service/product, request an upgrade, accept an offer etc…
+
+Import this notebook into your Databricks workspace to run this model for yourself (see how to [Import a notebook](https://docs.databricks.com/notebooks/notebooks-manage.html#import-a-notebook)).
+
+{{< jupyter databricks_training_model 500 >}}
