+++ content/predictive_ml_models/Databricks/using_your_model.md
@@ -1,0 +1,12 @@
++++
+title = "Using your Model"
+weight = 2
++++
+
+#### Predicting user's propensity to convert
+
+Now that your model is built and saved in MLflow registry, here is how you might load and use your model in a Databricks notebook.
+
+{{< jupyter databricks_deploying_model 500 >}}
+
+Now you are ready to use your model output for activation via Hightouch. The next chapter, [Data Activation]({{< ref data_activation >}}), guides you through how to target high propensity visitors to optimize ad spend.
